requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
google-generativeai==0.7.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pymupdf==1.24.9
pdfplumber==0.11.4
streamlit==1.38.0
python-dotenv==1.0.1
